Move getNoteName() to utils.js
This is self-contained at the moment, and useful for debugging elsewhere, if nothing else.
